+++ sdk/core/azure-core-amqp/src/models/amqp_value.cpp
@@ -103,6 +103,15 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
         case AMQP_TYPE_DOUBLE:
           os << "AMQP_TYPE_DOUBLE";
           break;
+        case AMQP_TYPE_DECIMAL128:
+          os << "AMQP_TYPE_DECIMAL128";
+          break;
+        case AMQP_TYPE_DECIMAL64:
+          os << "AMQP_TYPE_DECIMAL64";
+          break;
+        case AMQP_TYPE_DECIMAL32:
+          os << "AMQP_TYPE_DECIMAL32";
+          break;
         case AMQP_TYPE_CHAR:
           os << "AMQP_TYPE_CHAR";
           break;
@@ -178,6 +187,15 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
           break;
         case RustAmqpValueType::AmqpValueDouble:
           os << "AMQP_TYPE_DOUBLE";
+          break;
+        case RustAmqpValueType::AmqpValueDecimal32:
+          os << "AMQP_TYPE_DECIMAL32";
+          break;
+        case RustAmqpValueType::AmqpValueDecimal64:
+          os << "AMQP_TYPE_DECIMAL64";
+          break;
+        case RustAmqpValueType::AmqpValueDecimal128:
+          os << "AMQP_TYPE_DECIMAL128";
           break;
         case RustAmqpValueType::AmqpValueChar:
           os << "AMQP_TYPE_CHAR";
@@ -283,6 +301,15 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
       case AmqpValueType::Double:
         os << "Double";
         break;
+      case AmqpValueType::Decimal32:
+        os << "Decimal32";
+        break;
+      case AmqpValueType::Decimal64:
+        os << "Decimal64";
+        break;
+      case AmqpValueType::Decimal128:
+        os << "Decimal128";
+        break;
       case AmqpValueType::Char:
         os << "Char";
         break;
@@ -386,6 +413,23 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
           _detail::UniqueAmqpValueHandle{amqpvalue_create_double(value)})}
   {
   }
+
+  AmqpValue::AmqpValue(AmqpDecimal128 const& value)
+      : m_impl{std::make_unique<_detail::AmqpValueImpl>(
+          _detail::UniqueAmqpValueHandle{amqpvalue_create_decimal128(value.AsArray().data())})}
+  {
+  }
+  AmqpValue::AmqpValue(AmqpDecimal64 const& value)
+      : m_impl{std::make_unique<_detail::AmqpValueImpl>(
+          _detail::UniqueAmqpValueHandle{amqpvalue_create_decimal64(value.AsArray().data())})}
+  {
+  }
+  AmqpValue::AmqpValue(AmqpDecimal32 const& value)
+      : m_impl{std::make_unique<_detail::AmqpValueImpl>(
+          _detail::UniqueAmqpValueHandle{amqpvalue_create_decimal32(value.AsArray().data())})}
+  {
+  }
+
   AmqpValue::AmqpValue(char32_t value)
       : m_impl{std::make_unique<_detail::AmqpValueImpl>(
           _detail::UniqueAmqpValueHandle{amqpvalue_create_char(value)})}
@@ -578,6 +622,36 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
     return value;
   }
 
+  AmqpValue::operator AmqpDecimal32() const
+  {
+    std::array<uint8_t, 4> value{};
+    if (amqpvalue_get_decimal32(*m_impl, const_cast<uint8_t*>(value.data())))
+    {
+      throw std::runtime_error("Could not retrieve decimal32 value");
+    }
+    return AmqpDecimal32(value);
+  }
+
+  AmqpValue::operator AmqpDecimal64() const
+  {
+    std::array<uint8_t, 8> value{};
+    if (amqpvalue_get_decimal64(*m_impl, const_cast<uint8_t*>(value.data())))
+    {
+      throw std::runtime_error("Could not retrieve decimal64 value");
+    }
+    return AmqpDecimal64(value);
+  }
+
+  AmqpValue::operator AmqpDecimal128() const
+  {
+    std::array<uint8_t, 16> value{};
+    if (amqpvalue_get_decimal128(*m_impl, const_cast<uint8_t*>(value.data())))
+    {
+      throw std::runtime_error("Could not retrieve decimal128 value");
+    }
+    return AmqpDecimal128{value};
+  }
+
   AmqpValue::operator char32_t() const
   {
     std::uint32_t value = {};
@@ -753,6 +827,9 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
         {AMQP_TYPE_LONG, AmqpValueType::Long},
         {AMQP_TYPE_FLOAT, AmqpValueType::Float},
         {AMQP_TYPE_DOUBLE, AmqpValueType::Double},
+        {AMQP_TYPE_DECIMAL128, AmqpValueType::Decimal128},
+        {AMQP_TYPE_DECIMAL64, AmqpValueType::Decimal64},
+        {AMQP_TYPE_DECIMAL32, AmqpValueType::Decimal32},
         {AMQP_TYPE_CHAR, AmqpValueType::Char},
         {AMQP_TYPE_TIMESTAMP, AmqpValueType::Timestamp},
         {AMQP_TYPE_UUID, AmqpValueType::Uuid},
@@ -781,6 +858,9 @@ namespace Azure { namespace Core { namespace Amqp { namespace Models {
         {RustAmqpValueType::AmqpValueLong, AmqpValueType::Long},
         {RustAmqpValueType::AmqpValueFloat, AmqpValueType::Float},
         {RustAmqpValueType::AmqpValueDouble, AmqpValueType::Double},
+        {RustAmqpValueType::AmqpValueDecimal128, AmqpValueType::Decimal128},
+        {RustAmqpValueType::AmqpValueDecimal64, AmqpValueType::Decimal64},
+        {RustAmqpValueType::AmqpValueDecimal32, AmqpValueType::Decimal32},
         {RustAmqpValueType::AmqpValueChar, AmqpValueType::Char},
         {RustAmqpValueType::AmqpValueTimestamp, AmqpValueType::Timestamp},
         {RustAmqpValueType::AmqpValueUuid, AmqpValueType::Uuid},